+++ __mocks__/factories.ts
@@ -1,6 +1,8 @@
 import { createClient } from '@/datx/create-client';
-import { createFactory, sequence } from '@datx/test-data-factory';
 import { Flight } from '@/models/Flight';
+import { Session } from '@/models/Session';
+import { User } from '@/models/User';
+import { createFactory, oneOf, perBuild, sequence } from '@datx/test-data-factory';
 
 const client = createClient();
 const factory = createFactory(client);
@@ -14,5 +16,21 @@ export const flightFactory = factory(Flight, {
 		basePrice: '100',
 		currentSeatPrice: '100',
 		departsAt: '2021-01-01T00:00:00.000Z',
+	},
+});
+
+export const userFactory = factory(User, {
+	fields: {
+		id: sequence(),
+		firstName: oneOf('Johnny', 'Michael', 'Philip', 'Darius', 'Nieves', 'Christian'),
+		lastName: oneOf('Smith', 'Brown', 'Miller'),
+	},
+});
+
+export const sessionFactory = factory(Session, {
+	fields: {
+		id: sequence(),
+		email: sequence((x) => `user${x}@example.com`),
+		user: perBuild(() => userFactory()),
 	},
 });
