+++ public/locales/de-DE/flight-card.json
@@ -1,0 +1,3 @@
+{
+	"seatPriceLabel": "Sitzplatzpreis"
+}
