+++ public/locales/en-US/flight-card.json
@@ -1,0 +1,3 @@
+{
+	"seatPriceLabel": "Seat price"
+}
