+++ src/components/core/Card/Card.test.tsx
@@ -1,7 +1,7 @@
 import { axe } from 'jest-axe';
 
-import { Card } from './Card';
 import { render } from 'test-utils';
+import { Card } from './Card';
 
 describe('Card', () => {
 	it('should matches snapshot', () => {
