+++ src/components/shared/navigations/MainNavigation/components/UserMenu.tsx
@@ -1,6 +1,6 @@
 import { User } from '@/models/User';
 import { Avatar, forwardRef, IconButton, Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
+import { useTranslation } from 'next-i18next';
 
 export interface IUserMenuProps {
 	user?: User;
