+++ typings/i18next.d.ts
@@ -3,6 +3,7 @@
 import 'i18next';
 
 import common from 'public/locales/en-US/common.json';
+import flightCard from 'public/locales/en-US/flight-card.json';
 import flightListSection from 'public/locales/en-US/flight-list-section.json';
 import loginForm from 'public/locales/en-US/login-form.json';
 import login from 'public/locales/en-US/login.json';
@@ -16,6 +17,7 @@ declare module 'i18next' {
 		// custom resources type
 		resources: {
 			common: typeof common;
+			['flight-card']: typeof flightCard;
 			['flight-list-section']: typeof flightListSection;
 			['login-form']: typeof loginForm;
 			login: typeof login;
